+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 * Fix `ContextCapture::NoCapture` excluding popup config.
 * Add `ResponseVar::map_response`.
 * Add `Dialog!` widget, `DIALOG` service and related types.
-* *Deprecated* ` zng::task::http::get_text` and `Var::get_text`.
+* *Deprecated* `http::get_text`, ` http::Client::get_text` and `Var::get_text`.
     - Renamed to `get_txt`.
 * *Deprecated* `zng::window::native_dialog` module.
     - The new `zng::dialog` is the new surface API for all dialogs.